requirements.txt
pyyaml>=6.0
python-hcl2>=4.3
jsonschema>=4.17
